refactor(step5): Winners 객체를 cars의 우승자 결정 메소드에서 반환하도록 수정
